Валидация полей ввода - исправление архитектуры
