+++ js/server-requests.js
@@ -1,34 +1,7 @@
-import {makeMarkers, adForm, mapFilter} from './map.js';
-import {showMessage, closeMessageByEsc, closeMessageByClick, getOffDisabled} from './utils.js';
-import {clearForm, checkCorrectChoice} from './user-forms.js';
+import {makeMarkers, mapFilter} from './map.js';
+import {showMessage, getOffDisabled} from './utils.js';
 
-const mainContainer = document.querySelector('main');
 const SIMILAR_ANNOUNCEMENT_COUNT = 10;
-
-const showSuccessMessage = () => {
-  const successTemplate = document.querySelector('#success').content;
-  const successMessageContainer = successTemplate.querySelector('.success');
-  const successMessage = successMessageContainer.cloneNode(true);
-  successMessage.style.zIndex = 1000;
-  mainContainer.append(successMessage);
-  closeMessageByEsc(successMessage);
-  closeMessageByClick(successMessage);
-};
-
-const showErrorMessage = () => {
-  const errorTemplate = document.querySelector('#error').content;
-  const errorMessageContainer = errorTemplate.querySelector('.error');
-  const errorMessage = errorMessageContainer.cloneNode(true);
-  const errorButton = errorMessage.querySelector('.error__button');
-  errorMessage.style.zIndex = 1000;
-  mainContainer.append(errorMessage);
-
-  errorButton.addEventListener('click', () => {
-    errorMessage.remove();
-  });
-  closeMessageByEsc(errorMessage);
-  closeMessageByClick(errorMessage);
-};
 
 let announcementsArray;
 
@@ -45,30 +18,24 @@ fetch('https://22.javascript.pages.academy/keksobooking/data')
     }
   });
 
-adForm.addEventListener('submit', (evt) => {
-  evt.preventDefault();
-  if (checkCorrectChoice()){
-    const formData = new FormData(evt.target);
-    fetch(
-      'https://22.javascript.pages.academy/keksobooking',
-      {
-        method: 'POST',
-        body: formData,
-      },
-    )
-      .then((response) => {
-        if(response.ok){
-          showSuccessMessage();
-          clearForm();
-        } else {
-          showErrorMessage();
-        }
+const sendData = (onSuccess, onFail, data) => {
+  fetch(
+    'https://22.javascript.pages.academy/keksobooking',
+    {
+      method: 'POST',
+      body: data,
+    },
+  )
+    .then((response) => {
+      if(response.ok){
+        onSuccess();
+      } else {
+        onFail();
+      }
+    })
+    .catch(() => {
+      onFail();
+    })
+};
 
-      })
-      .catch(() => {
-        showErrorMessage();
-      })
-  }
-});
-
-export {announcementsArray};
+export {announcementsArray, sendData};
